+++ twister2/stockanalysis/src/main/java/edu/iu/dsc/tws/apps/stockanalysis/MDSWorkerComputeTask.java
@@ -1,7 +1,7 @@
 package edu.iu.dsc.tws.apps.stockanalysis;
 
-import edu.iu.dsc.tws.api.task.IMessage;
-import edu.iu.dsc.tws.api.task.nodes.BaseCompute;
+import edu.iu.dsc.tws.api.compute.IMessage;
+import edu.iu.dsc.tws.api.compute.nodes.BaseCompute;
 
 import java.util.List;
 import java.util.logging.Logger;
