+++ twister2/stockanalysis/src/main/java/edu/iu/dsc/tws/apps/stockanalysis/StockAnalysisSinkTask.java
@@ -1,13 +1,14 @@
 package edu.iu.dsc.tws.apps.stockanalysis;
 
+import edu.iu.dsc.tws.api.compute.IMessage;
+import edu.iu.dsc.tws.api.compute.TaskContext;
+import edu.iu.dsc.tws.api.compute.modifiers.Collector;
+import edu.iu.dsc.tws.api.compute.nodes.BaseSink;
 import edu.iu.dsc.tws.api.config.Config;
 import edu.iu.dsc.tws.api.dataset.DataPartition;
-import edu.iu.dsc.tws.api.task.IMessage;
-import edu.iu.dsc.tws.api.task.TaskContext;
-import edu.iu.dsc.tws.api.task.modifiers.Collector;
-import edu.iu.dsc.tws.api.task.nodes.BaseSink;
-import edu.iu.dsc.tws.dataset.impl.EntityPartition;
+import edu.iu.dsc.tws.dataset.partition.EntityPartition;
 
+import java.util.Set;
 import java.util.logging.Logger;
 
 public class StockAnalysisSinkTask extends BaseSink implements Collector {
@@ -27,5 +28,10 @@ public class StockAnalysisSinkTask extends BaseSink implements Collector {
     @Override
     public DataPartition<double[][]> get() {
         return new EntityPartition<>(context.taskIndex(), null);
+    }
+
+    @Override
+    public Set<String> getCollectibleNames() {
+        return null;
     }
 }
